refactor: extract cloud particle specific logic to `cloud_particle.c`
